fix(meeting-bot): break node host import cycle
